+++ src/markets/forex.rs
@@ -1,0 +1,1 @@
+pub mod quotes;
