+++ src/markets/forex/quotes/forwardpoints.rs
