--- src/markets/ir/cashrates/sofr.rs
@@ -1,1 +1,0 @@
-
