+++ src/time.rs
@@ -1,5 +1,4 @@
 pub mod businessdayconvention;
 pub mod calendars;
-pub mod date;
 pub mod daycounters;
 pub mod period;
