+++ src/time/calendars.rs
@@ -1,5 +1,6 @@
 use crate::time::businessdayconvention::BusinessDayConvention;
 use chrono::{Datelike, Duration, NaiveDate, Weekday};
+use serde::Serialize;
 
 static EASTER_MONDAY: [u32; 299] = [
     98, 90, 103, 95, 114, 106, 91, 111, 102, // 1901-1909
@@ -67,7 +68,6 @@ static ORTHODOX_EASTER_MONDAY: [u32; 299] = [
     116, 108, 128, 119, 104, 124, 116, 100, 120, 112, // 2190-2199
 ];
 
-pub mod all;
 pub mod argentina;
 pub use argentina::Argentina;
 pub mod austria;
