+++ src/time/calendars/australia.rs
@@ -1,5 +1,6 @@
 // Holidays in Australia.
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 pub enum AustraliaMarket {
