+++ src/time/calendars/austria.rs
@@ -1,6 +1,7 @@
 // Holidays in Austria.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 pub enum AustriaMarket {
@@ -99,6 +100,7 @@ impl Calendar for Austria {
 mod tests {
     use super::Austria;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
