+++ src/time/calendars/botswana.rs
@@ -1,6 +1,7 @@
 // Holidays in Botswana.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -51,6 +52,7 @@ impl Calendar for Botswana {
 mod tests {
     use super::Botswana;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
