+++ src/time/calendars/canada.rs
@@ -1,6 +1,7 @@
 // Holidays in Canada.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 pub enum CanadaMarket {
@@ -80,6 +81,7 @@ impl Calendar for Canada {
 mod tests {
     use super::Canada;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
