+++ src/time/calendars/china.rs
@@ -1,5 +1,6 @@
 // Holidays in China.
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -155,6 +156,7 @@ impl Calendar for China {
 mod tests {
     use super::China;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
