+++ src/time/calendars/czechrepublic.rs
@@ -1,6 +1,7 @@
 // Holidays in Czech Republic.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -53,6 +54,7 @@ impl Calendar for CzechRepublic {
 mod tests {
     use super::CzechRepublic;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
