+++ src/time/calendars/france.rs
@@ -1,6 +1,7 @@
 // Holidays in France.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 pub enum FranceMarket {
@@ -90,6 +91,7 @@ impl Calendar for France {
 mod tests {
     use super::France;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
