+++ src/time/calendars/hungary.rs
@@ -1,6 +1,7 @@
 // Holidays in Hungary.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -49,6 +50,7 @@ impl Calendar for Hungary {
 mod tests {
     use super::Hungary;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
