+++ src/time/calendars/iceland.rs
@@ -1,6 +1,7 @@
 // Holidays in Iceland.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -47,6 +48,7 @@ impl Calendar for Iceland {
 mod tests {
     use super::Iceland;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
