+++ src/time/calendars/indonesia.rs
@@ -1,6 +1,7 @@
 // Holidays in Indonesia.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -231,6 +232,7 @@ impl Calendar for Indonesia {
 mod tests {
     use super::Indonesia;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
