+++ src/time/calendars/israel.rs
@@ -1,6 +1,7 @@
 // Holidays in Israel.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -320,6 +321,7 @@ impl Calendar for Israel {
 mod tests {
     use super::Israel;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
