+++ src/time/calendars/italy.rs
@@ -1,6 +1,7 @@
 // Holidays in Italy.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 pub enum ItalyMarket {
@@ -92,6 +93,7 @@ impl Calendar for Italy {
 mod tests {
     use super::Italy;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
