+++ src/time/calendars/japan.rs
@@ -1,6 +1,7 @@
 // Holidays in Japan.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -134,6 +135,7 @@ impl Calendar for Japan {
 mod tests {
     use super::Japan;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
