+++ src/time/calendars/jointcalendar.rs
@@ -1,4 +1,5 @@
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
