+++ src/time/calendars/newzealand.rs
@@ -1,6 +1,7 @@
 // Holidays in New Zealand.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -72,6 +73,7 @@ impl Calendar for NewZealand {
 mod tests {
     use super::NewZealand;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
