+++ src/time/calendars/singapore.rs
@@ -1,6 +1,7 @@
 // Holidays in Singapore.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -189,6 +190,7 @@ impl Calendar for Singapore {
 mod tests {
     use super::Singapore;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
