+++ src/time/calendars/slovakia.rs
@@ -1,6 +1,7 @@
 // Holidays in Slovakia.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -56,6 +57,7 @@ impl Calendar for Slovakia {
 mod tests {
     use super::Slovakia;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
