+++ src/time/calendars/southkorea.rs
@@ -1,6 +1,7 @@
 // Holidays in South Korea.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -228,6 +229,7 @@ impl Calendar for SouthKorea {
 mod tests {
     use super::SouthKorea;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
