+++ src/time/calendars/sweden.rs
@@ -1,6 +1,7 @@
 // Holidays in Sweden.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::{NaiveDate, Weekday};
 
 #[derive(Default)]
@@ -50,6 +51,7 @@ impl Calendar for Sweden {
 mod tests {
     use super::Sweden;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
