+++ src/time/calendars/switzerland.rs
@@ -1,6 +1,7 @@
 // Holidays in Switzerland.
 
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -43,6 +44,7 @@ impl Calendar for Switzerland {
 mod tests {
     use super::Switzerland;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
