+++ src/time/calendars/target.rs
@@ -1,5 +1,6 @@
 // Holidays - Target.
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -37,6 +38,7 @@ impl Calendar for Target {
 mod tests {
     use super::Target;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Duration, NaiveDate};
 
     #[test]
