+++ src/time/calendars/weekendsonly.rs
@@ -1,5 +1,6 @@
 // Holidays for weekends-only.
 use crate::time::calendars::Calendar;
+use serde::Serialize;
 use chrono::NaiveDate;
 
 #[derive(Default)]
@@ -15,6 +16,7 @@ impl Calendar for WeekendsOnly {
 mod tests {
     use super::WeekendsOnly;
     use crate::time::calendars::Calendar;
+use serde::Serialize;
     use chrono::{Datelike, Duration, NaiveDate, Weekday};
 
     #[test]
